+++ NEWS.md
@@ -4,6 +4,8 @@
 * Added checking input functions.
 * Added stan functions for primary censored and truncated distributions.
 * Added R functions for primary censored and truncated distributions.
+* Add R function to facilitate working with the Stan code.
 * Added tests for primary censored and truncated distributions.
 * Added tests to compare R and Stan implementations.
+* Added tests for the R functions that facilitate working with the Stan code.
 * Resolved R CMD check errors, warnings and notes.